# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PillPal Backend API",
    "version": "1.0.0",
    "description": "PillPal API docs generated from endpoint annotations."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/medicine-lookups": {
      "get": {
        "tags": [
          "Medicine Lookup"
        ],
        "summary": "List current user's unknown medicine lookup attempts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "in_progress",
                "needs_admin_review",
                "verified",
                "rejected",
                "failed"
              ]
            }
          },
          {
            "in": "query",
            "name": "queryName",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/medicine-lookups/sources": {
      "get": {
        "tags": [
          "Medicine Lookup"
        ],
        "summary": "List configured medicine data sources",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "sourceType",
            "schema": {
              "type": "string",
              "enum": [
                "distributor",
                "administration",
                "general_web"
              ]
            }
          },
          {
            "in": "query",
            "name": "active",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/medicine-lookups/{id}/run": {
      "post": {
        "tags": [
          "Medicine Lookup"
        ],
        "summary": "Run the unknown medicine lookup orchestrator",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Medicine lookup not found"
          }
        }
      }
    },
    "/medicine-lookups/{id}/save-medication": {
      "post": {
        "tags": [
          "Medicine Lookup"
        ],
        "summary": "Save a verified external lookup candidate as a user medication",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "externalCandidateId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "note": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Lookup is not verified or has no authorized candidate"
          },
          "404": {
            "description": "Medicine lookup not found"
          }
        }
      }
    },
    "/medicine-lookups/{id}": {
      "get": {
        "tags": [
          "Medicine Lookup"
        ],
        "summary": "Get current user's medicine lookup attempt detail",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Medicine lookup not found"
          }
        }
      }
    },
...
  },
...
}